Webpack backend together with frontend in one package.json => Unify whole app + Minify server.js + Add concurrently to monitor both front and back ends
