+++ frontend/components/Footer.scss
