+++ frontend/components/Home.scss
