+++ frontend/components/NavigationBar.scss
