--- frontend/package.json
@@ -1,47 +1,0 @@
-{
-  "name": "frontend",
-  "version": "1.0.0",
-  "description": "MERN App Frontend",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "prebuild": "rimraf ./dist/*",
-    "build": "webpack --config=webpack.config.prod.js",
-    "start": "webpack-dev-server --open --config=webpack.config.dev.js"
-  },
-  "author": "DBC",
-  "license": "ISC",
-  "devDependencies": {
-    "@types/react": "^16.7.18",
-    "@types/react-dom": "^16.0.11",
-    "@types/react-router": "^4.4.3",
-    "@types/react-router-dom": "^4.3.1",
-    "css-loader": "^2.1.0",
-    "file-loader": "^3.0.1",
-    "fork-ts-checker-notifier-webpack-plugin": "^1.0.2",
-    "fork-ts-checker-webpack-plugin": "^1.3.7",
-    "glob": "^7.1.3",
-    "html-webpack-plugin": "^3.2.0",
-    "imagemin-webpack-plugin": "^2.4.2",
-    "mini-css-extract-plugin": "^0.5.0",
-    "node-sass": "^4.12.0",
-    "optimize-css-assets-webpack-plugin": "^5.0.1",
-    "rimraf": "^2.6.3",
-    "sass-loader": "^7.1.0",
-    "source-map-loader": "^0.2.4",
-    "style-loader": "^0.23.1",
-    "ts-loader": "^5.3.2",
-    "tslint": "^5.18.0",
-    "typescript": "^3.2.2",
-    "uglifyjs-webpack-plugin": "^2.1.1",
-    "webpack": "^4.28.3",
-    "webpack-cli": "^3.1.2",
-    "webpack-dev-server": "^3.1.14"
-  },
-  "dependencies": {
-    "react": "^16.7.0",
-    "react-dom": "^16.7.0",
-    "react-router": "^4.3.1",
-    "react-router-dom": "^4.3.1"
-  }
-}
