--- frontend/tsconfig.json
@@ -1,22 +1,0 @@
-{
-    "compilerOptions": {
-        "outDir": "./dist",
-        "sourceMap": true,
-        "noImplicitAny": false,
-        "removeComments": true,
-        "module": "commonjs",
-        "target": "es5",
-        "jsx": "react",
-        "lib": [
-            "dom",
-            "es7"
-        ]
-    },
-    "include": [
-        "./src/**/*"
-    ],
-    "typeRoots": [
-        "../node_modules/@types",
-        "../@types"
-    ]
-}
