--- frontend/tslint.json
@@ -1,89 +1,0 @@
-{
-    "extends": [
-        "tslint:recommended"
-    ],
-    "rulesDirectory": ["node_modules/tslint-microsoft-contrib"],
-    "rules": {
-        "no-console": false,
-        "class-name": true,
-        "comment-format": [
-            true,
-            "check-space"
-        ],
-        "indent": [
-            true,
-            "spaces"
-        ],
-        "one-line": [
-            true,
-            "check-open-brace",
-            "check-whitespace"
-        ],
-        "quotemark": [
-            true,
-            "double",
-            "avoid-escape"
-        ],
-        "semicolon": [
-            true,
-            "always",
-            "ignore-bound-class-methods"
-        ],
-        "whitespace": [
-            true,
-            "check-branch",
-            "check-decl",
-            "check-operator",
-            "check-module",
-            "check-separator",
-            "check-type"
-        ],
-        "typedef-whitespace": [
-            true,
-            {
-                "call-signature": "nospace",
-                "index-signature": "nospace",
-                "parameter": "nospace",
-                "property-declaration": "nospace",
-                "variable-declaration": "nospace"
-            },
-            {
-                "call-signature": "onespace",
-                "index-signature": "onespace",
-                "parameter": "onespace",
-                "property-declaration": "onespace",
-                "variable-declaration": "onespace"
-            }
-        ],
-        "no-var-requires": false,
-        "no-var-keyword": true,
-        "no-internal-module": true,
-        "no-trailing-whitespace": true,
-        "no-null-keyword": true,
-        "prefer-const": true,
-        "jsdoc-format": true,
-        "no-implicit-dependencies": [
-            true,
-            "dev"
-        ],
-        "no-string-literal": false,
-        "only-arrow-functions": false,
-        "space-before-function-paren": [
-            true,
-            {
-                "anonymous": "always",
-                "named": "never",
-                "asyncArrow": "always"
-            }
-        ],
-        "object-literal-sort-keys": false,
-        "forin": false,
-        "ordered-imports": false,
-        "member-ordering": false,
-        "max-classes-per-file": false,
-        "trailing-comma": false,
-        "object-literal-shorthand": false,
-        "max-func-body-length": [true, 50, {"ignore-comments": true, "ignore-parameters-to-function-regex": "^describe$" }],
-        "max-file-line-count": [true, 3000]
-    }
-}
