+++ frontend/types/staticAssets.d.ts
