+++ package.json
@@ -1,25 +1,21 @@
 {
-    "name": "backend",
+    "name": "full-stack",
     "version": "1.0.0",
-    "description": "MERN App Backend",
+    "description": "T-MERN App",
     "main": "dist/server.js",
     "scripts": {
-        "build-prod": "npm run prep-frontend && npm run prep-backend",
+        "build-prod": "npm install && npm run build && ts-node deleteDev.ts && rimraf deleteDev.ts && npm prune --production",
         "heroku-postbuild": "npm run build-prod",
-        "prep-frontend": "cd frontend && npm install && npm run build",
-        "prep-backend": "npm install && npm run build && npm run delete-dev && rimraf deleteDev.ts && npm prune --production",
-        "install-frontend": "cd frontend && npm install",
-        "build-frontend": "cd frontend && npm run build",
-        "delete-dev": "ts-node deleteDev.ts",
         "prebuild": "rimraf ./dist/*",
-        "build": "npm run build-ts && npm run tslint && npm run copy-dist-frontend",
-        "build-ts": "tsc",
-        "tslint": "tslint -c tslint.json -p tsconfig.json",
-        "copy-dist-frontend": "ts-node copyFrontend.ts",
+        "build": "webpack --config=webpack.config.prod.js",
+        "build-dev": "npm run prebuild && webpack --config=webpack.config.dev.js",
         "start": "npm run serve",
         "serve": "node dist/server.js",
-        "watch-node": "nodemon dist/server.js",
-        "watch": "concurrently -k -p \"[{name}]\" -n \"Sass,TypeScript,Node\" -c \"yellow.bold,cyan.bold,green.bold\" \"npm run watch-sass\" \"npm run watch-ts\" \"npm run watch-node\"",
+        "watch-frontend": "webpack-dev-server --open --stack=frontend --config=webpack.config.dev.js",
+        "watch-backend": "nodemon dist/server.js",
+        "watch": "concurrently -k -p \"[{name}]\" -n \"React,Node\" -c \"yellow.bold,cyan.bold\" \"npm run watch-frontend\" \"npm run watch-backend\"",
+        "build-ts": "tsc",
+        "tslint": "tslint -c tslint.json -p tsconfig.json",
         "test": "jest --forceExit --coverage --verbose",
         "watch-test": "npm run test -- --watchAll",
         "watch-ts": "tsc -w",
@@ -27,7 +23,7 @@
         "serve-debug": "nodemon --inspect dist/server.js",
         "watch-debug": "concurrently -k -p \"[{name}]\" -n \"Sass,TypeScript,Node\" -c \"yellow.bold,cyan.bold,green.bold\" \"npm run watch-sass\" \"npm run watch-ts\" \"npm run serve-debug\""
     },
-    "keywords": [],
+    "keywords": ["Typescript", "MongoDB", "ExpressJS", "React", "NodeJS"],
     "author": "DBC",
     "license": "ISC",
     "dependencies": {
@@ -48,6 +44,10 @@
         "mongoose": "^5.6.0",
         "passport": "^0.4.0",
         "path": "^0.12.7",
+        "react": "^16.7.0",
+        "react-dom": "^16.7.0",
+        "react-router": "^4.3.1",
+        "react-router-dom": "^4.3.1",
         "winston": "^3.2.1"
     },
     "devDependencies": {
@@ -65,16 +65,39 @@
         "@types/mongodb": "^3.1.28",
         "@types/mongoose": "^5.5.6",
         "@types/passport": "^1.0.0",
+        "@types/react": "^16.7.18",
+        "@types/react-dom": "^16.0.11",
+        "@types/react-router": "^4.4.3",
+        "@types/react-router-dom": "^4.3.1",
         "@types/shelljs": "^0.8.5",
         "@types/winston": "^2.4.4",
+        "concurrently": "^4.1.1",
+        "css-loader": "^2.1.0",
+        "file-loader": "^3.0.1",
+        "fork-ts-checker-notifier-webpack-plugin": "^1.0.2",
+        "fork-ts-checker-webpack-plugin": "^1.3.7",
         "glob": "^7.1.4",
+        "html-webpack-plugin": "^3.2.0",
+        "imagemin-webpack-plugin": "^2.4.2",
+        "mini-css-extract-plugin": "^0.5.0",
+        "node-sass": "^4.12.0",
         "nodemon": "^1.19.1",
+        "optimize-css-assets-webpack-plugin": "^5.0.1",
         "rimraf": "^2.6.3",
+        "sass-loader": "^7.1.0",
         "shelljs": "^0.8.3",
+        "source-map-loader": "^0.2.4",
+        "style-loader": "^0.23.1",
+        "ts-loader": "^5.3.2",
         "ts-node": "^8.3.0",
         "tslint": "^5.18.0",
         "tslint-microsoft-contrib": "^6.2.0",
-        "typescript": "^3.5.2"
+        "typescript": "^3.5.2",
+        "uglifyjs-webpack-plugin": "^2.1.1",
+        "webpack": "^4.28.3",
+        "webpack-cli": "^3.1.2",
+        "webpack-dev-server": "^3.1.14",
+        "webpack-node-externals": "^1.7.2"
     },
     "engines": {
         "node": "10.16.0",
