+++ tsconfig.json
@@ -8,6 +8,7 @@
         "emitDecoratorMetadata": true,
         "experimentalDecorators": true,
         "moduleResolution": "node",
+        "jsx": "react",
         "target": "es5",
         "lib": [
             "dom",
@@ -22,6 +23,7 @@
         }
     },
     "include": [
-        "./src/**/*"
+        "./frontend/**/*",
+        "./src/**/*",
     ],
 }
